Added UI elements
Started by adding the TMPro package, added a settings page, added a player input system which will need to be connected to the movement script in order to use the rebinding function. I made it really simple.
Also added a basic menu page with a start game which will need to be linked to the game scene.